+++ test/iobuffer.jl
@@ -49,7 +49,7 @@ end
 
 let io = IOBuffer("hamster\nguinea pig\nturtle")
 @test position(io) == 0
-@test readline(io) == "hamster\n"
+@test readline(io, false) == "hamster\n"
 @test readstring(io) == "guinea pig\nturtle"
 @test_throws EOFError read(io,UInt8)
 seek(io,0)
@@ -66,14 +66,28 @@ let io = PipeBuffer()
 @test_throws EOFError read(io,UInt8)
 @test write(io,"pancakes\nwaffles\nblueberries\n") > 0
 @test position(io) == 0
-@test readline(io) == "pancakes\n"
+@test readline(io, false) == "pancakes\n"
 Base.compact(io)
-@test readline(io) == "waffles\n"
+@test readline(io, false) == "waffles\n"
 @test write(io,"whipped cream\n") > 0
-@test readline(io) == "blueberries\n"
+@test readline(io, false) == "blueberries\n"
 @test_throws ArgumentError seek(io,0)
 @test_throws ArgumentError truncate(io,0)
-@test readline(io) == "whipped cream\n"
+@test readline(io, false) == "whipped cream\n"
+@test write(io,"pancakes\nwaffles\nblueberries\n") > 0
+@test readlines(io, true) == String["pancakes", "waffles", "blueberries"]
+write(io,"\n\r\n\n\r \n") > 0
+@test readlines(io, false) == String["\n", "\r\n", "\n", "\r \n"]
+write(io,"\n\r\n\n\r \n") > 0
+@test readlines(io, true) == String["", "", "", "\r "]
+@test write(io,"α\nβ\nγ\nδ") > 0
+@test readlines(io, true) == String["α", "β", "γ", "δ"]
+@test write(io,"α\nβ\nγ\nδ") > 0
+@test readlines(io, false) == String["α\n","β\n","γ\n","δ"]
+@test readlines(IOBuffer(""), false) == []
+@test readlines(IOBuffer(""), true) == []
+@test readlines(IOBuffer("first\nsecond"), false) == String["first\n", "second"]
+@test readlines(IOBuffer("first\nsecond"), true) == String["first", "second"]
 Base.compact(io)
 @test position(io) == 0
 @test ioslength(io) == 0
@@ -133,7 +147,7 @@ a = Array{UInt8}(1024)
 @test eof(io)
 end
 
-@test isempty(readlines(IOBuffer()))
+@test isempty(readlines(IOBuffer(), false))
 
 # issue #8193
 let io=IOBuffer("asdf")